+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `Log` file name.
 - Update `ricaun.NUnit` to `1.4.1`.
 - Add `MetadataMapper` to map `AssemblyMetadata` configuration.
+- Update to use `Task.Run` to force run outside the Revit context.
 ### TestAdapter
 - Update `MapperKey` to multiple converts.
 - Update `MetadataMapper` with `prefix`.