+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `ricaun.NUnit` to `1.4.1`.
 - Add `MetadataMapper` to map `AssemblyMetadata` configuration.
 - Update to use `Task.Run` to force run outside the Revit context.
+### Command
+- Update `AssemblyName` to use `Version`.
 ### TestAdapter
 - Update `MapperKey` to multiple converts.
 - Update `MetadataMapper` with `prefix`.