+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### TestAdapter
 - Update `MapperKey` to multiple converts.
 - Update `MetadataMapper` with `prefix`.
+- Fix `Mapper` convert integer to ignore dot.
 
 ## [1.7.0] / 2024-11-02 - 2024-12-03
 ### Features