+++ CHANGELOG.md
@@ -4,12 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.7.1] / 2024-12-07
+## [1.7.1] / 2024-12-07 - 2024-12-10
 ### Application
 - Update to use `AssemblyMetadata` to set configuration.
 - Update `TestExecuteUtils` to use `ConfigurationMetadata`.
 - Update `Log` file name.
-- Update `ricaun.NUnit` to `1.4.1-beta.1`.
+- Update `ricaun.NUnit` to `1.4.1-rc`.
 - Add `MetadataMapper` to map `AssemblyMetadata` configuration.
 - Update to use `Task.Run` to force run outside the Revit context.
 - Update `Timeout` to time unit in minutes.